requirements.txt
streamlit
sqlalchemy
pydantic
pydantic-settings
httpx
aiohttp
python-dotenv
openai
sqlalchemy-utils
pandas
openpyxl
gspread
oauth2client
tenacity
beautifulsoup4
firecrawl-py
google-search-results
rich
asyncio
duckduckgo-search
